pyproject.toml
[project]
name = "pravesh-backend"
version = "0.1.0"
description = "Working Backend for our PRAVESH the smart entry-exit system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "attrs>=25.3.0",
    "babel>=2.17.0",
    "fastapi>=0.116.0",
    "fastapi-admin>=1.0.4",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
